Updated Favicon code for non-supported
After recent discussion at work, I have updated Favicon code for non-supported browsers and Web Servers considering this project is open source and not everyone uses web servers or browsers that automatically recognize the favicon in the root directory. This change comes after it was previously removed but I feel it needs adding back as this is an open source project and it is always recommended for developers to add the extra line of code in anyways (not like its a huge deal but anyway).