Clarification on RX group.
Peach Host - WHMCS was never run under no license. This was an issue with a load balancer which resulted in Peach host having to contact WHMCS in regards to getting rid of IP restrictions which took some time. It would have helped if this was said in advance. As for downtime, from users experiences this as well as support is decently bad. 
As for RX Group, Etheral Solution, and FastSrv, I have been staff there in the past and been a member and buyer for many months and I see no issues nor complaints from anyone. If you want evidence to back any of this up head over to their discords. 
If this summer hosts idea is going to continue, it has to be done right and that means it has to be based off evidence and not just assumptions or saying things.